+++ react/ProductDescription.js
@@ -1,3 +1,0 @@
-import ProductDescription from './components/ProductDescription/Wrapper'
-
-export default ProductDescription