+++ react/SearchBar.js
@@ -1,3 +1,0 @@
-import SearchBar from './components/SearchBar/index'
-
-export default SearchBar