+++ react/SearchBar.tsx
@@ -1,18 +1,56 @@
 import React, { useState, useCallback } from 'react'
 import { useIntl } from 'react-intl'
-import PropTypes from 'prop-types'
 import { ModalContext } from 'vtex.modal-layout'
 import { useRuntime } from 'vtex.render-runtime'
 
-import SearchBar from './components/SearchBar'
-import encodeForwardSlash from '../../utils/encodeForwardSlash'
+import SearchBar from './components/SearchBar/SearchBar'
+import encodeForwardSlash from './utils/encodeForwardSlash'
 
 const { useModalDispatch } = ModalContext
+
+interface Props {
+  /** Define when use the compact version of the component */
+  compactMode?: boolean
+  /** Define if the search icon is on left or right position */
+  hasIconLeft?: boolean
+  /** Custom classes for the search icon */
+  iconClasses?: string
+  /** Define if the search input should autofocus or not */
+  autoFocus?: boolean
+  /** Max width of the search bar */
+  maxWidth?: string | number
+  /**
+   * Uses the term the user has inputed to try to navigate to the proper
+   * page type (e.g. a department, a brand, a category)
+   */
+  attemptPageTypeSearch?: boolean
+  /** A template for a custom url. It can have a substring ${term} used as placeholder to interpolate the searched term. (e.g. `/search?query=${term}`) */
+  customSearchPageUrl?: string
+  placeholder?: string
+  /* Autocomplete Horizontal alignment */
+  autocompleteAlignment?: 'right' | 'left' | 'center'
+  /** Define if autocomplete should be open on input focus or not */
+  openAutocompleteOnFocus?: boolean
+  /** Define if input should blur on submit */
+  blurOnSubmit?: boolean
+  /** Define if icon should submit on click */
+  submitOnIconClick?: boolean
+  /** Minimum search term length allowed */
+  minSearchTermLength?: number
+  /** If true, the autocomplete will fill the whole window horizontally */
+  autocompleteFullWidth?: boolean
+  /** The type of the search input */
+  inputType?: 'text' | 'search'
+  /** Define the component display mode, such as which buttons should be visible */
+  displayMode?: 'clear-button' | 'search-and-clear-buttons' | 'search-button'
+  /** Define how the autocomplete component should be displayed. Possible values are: `overlay` (suggestions overlapping other components) and `container` (displays the suggestion within a container). */
+  containerMode: 'overlay' | 'container'
+}
 
 /**
  * Canonical search bar that uses the autocomplete endpoint to search for a specific product
  * */
-function SearchBarContainer(props) {
+function SearchBarContainer(props: Props) {
   const intl = useIntl()
   const {
     compactMode,
@@ -27,6 +65,7 @@ function SearchBarContainer(props) {
     }),
     autocompleteAlignment = 'right',
     openAutocompleteOnFocus = false,
+    containerMode = 'overlay',
     blurOnSubmit = false,
     submitOnIconClick,
     displayMode = 'clear-button',
@@ -117,47 +156,13 @@ function SearchBarContainer(props) {
       minSearchTermLength={minSearchTermLength}
       autocompleteFullWidth={autocompleteFullWidth}
       inputType={inputType}
+      containerMode={containerMode}
     />
   )
 }
 
 SearchBarContainer.schema = {
   title: 'admin/editor.search-bar.title',
-}
-
-SearchBarContainer.propTypes = {
-  /** Indentify when use the compact version of the component */
-  compactMode: PropTypes.bool,
-  /** Identify if the search icon is on left or right position */
-  hasIconLeft: PropTypes.bool,
-  /** Custom classes for the search icon */
-  iconClasses: PropTypes.string,
-  /** Identify if the search input should autofocus or not */
-  autoFocus: PropTypes.bool,
-  /** Max width of the search bar */
-  maxWidth: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
-  /** Uses the term the user has inputed to try to navigate to the proper
-   * page type (e.g. a department, a brand, a category)
-   */
-  attemptPageTypeSearch: PropTypes.bool,
-  /** A template for a custom url. It can have a substring ${term} used as placeholder to interpolate the searched term. (e.g. `/search?query=${term}`) */
-  customSearchPageUrl: PropTypes.string,
-  placeholder: PropTypes.string,
-  /* Autocomplete Horizontal alignment */
-  autocompleteAlignment: PropTypes.string,
-  /** Identify if autocomplete should be open on input focus or not */
-  openAutocompleteOnFocus: PropTypes.bool,
-  /** Identify if input should blur on submit */
-  blurOnSubmit: PropTypes.bool,
-  /** Identify if icon should submit on click */
-  submitOnIconClick: PropTypes.bool,
-  displayMode: PropTypes.string,
-  /** Minimum search term length allowed */
-  minSearchTermLength: PropTypes.number,
-  /** If true, the autocomplete will fill the whole window horizontally */
-  autocompleteFullWidth: PropTypes.bool,
-  /** The type of the search input */
-  inputType: PropTypes.oneOf(['text', 'search']),
 }
 
 export default SearchBarContainer