+++ react/components/SearchBar/SearchBar.tsx
@@ -1,7 +1,6 @@
 import React, { useRef, useCallback, useState, useEffect, useMemo } from 'react'
-import PropTypes from 'prop-types'
 import classNames from 'classnames'
-import Downshift from 'downshift'
+import Downshift, { DownshiftProps } from 'downshift'
 import debounce from 'debounce'
 import {
   NoSSR,
@@ -13,13 +12,14 @@ import { Overlay } from 'vtex.react-portal'
 import { useCssHandles, applyModifiers } from 'vtex.css-handles'
 import { defineMessages, useIntl } from 'react-intl'
 
-import styles from '../styles.css'
+import styles from './SearchBar.css'
 import AutocompleteResults from '../../AutocompleteResults'
+import type { Props as AutocompleteResultsProps } from '../../AutocompleteResults'
 import AutocompleteInput from './AutocompleteInput'
 
 const CSS_HANDLES = ['searchBarInnerContainer']
 const SEARCH_DELAY_TIME = 1000
-const AUTCOMPLETE_EXTENSION_ID = 'autocomplete-result-list'
+const AUTOCOMPLETE_EXTENSION_ID = 'autocomplete-result-list'
 
 const messages = defineMessages({
   searchTermTooShort: {
@@ -28,7 +28,56 @@ const messages = defineMessages({
   },
 })
 
-const SearchBar = ({
+interface Props {
+  /** Placeholder to be used on the input */
+  placeholder: string
+  /** Current value of the input */
+  inputValue: string
+  /** Function to handle input changes */
+  onInputChange: DownshiftProps<any>['onChange']
+  /** Function to direct the user to the searchPage */
+  onGoToSearchPage: () => void
+  /** Function to clear the input */
+  onClearInput: () => void
+  /** Indentify when use the compact version of the component */
+  compactMode?: boolean
+  /** Identify if the search icon is on left or right position */
+  hasIconLeft?: boolean
+  /** Custom classes for the search icon */
+  iconClasses?: string
+  /** Block class for the search icon */
+  iconBlockClass?: string
+  /** Identify if the search input should autofocus or not */
+  autoFocus?: boolean
+  /** Max width of the search bar */
+  maxWidth?: string | number
+  /** A template for a custom url. It can have a substring ${term} used as placeholder to interpolate the searched term. (e.g. `/search?query=${term}`) */
+  customSearchPageUrl?: string
+  /** Uses the term the user has inputed to try to navigate to the proper
+   * page type (e.g. a department, a brand, a category)
+   */
+  attemptPageTypeSearch?: boolean
+  /* Autocomplete Horizontal alignment */
+  autocompleteAlignment?: 'right' | 'left' | 'center'
+  /** Identify if autocomplete should be open on input focus or not */
+  openAutocompleteOnFocus?: boolean
+  /** Identify if input should blur on submit */
+  blurOnSubmit?: boolean
+  /** Identify if icon should submit on click */
+  submitOnIconClick?: boolean
+  /** Minimum search term length allowed */
+  minSearchTermLength?: number
+  /** If true, the autocomplete will fill the whole window horizontally */
+  autocompleteFullWidth?: boolean
+  /** The type of the search input */
+  inputType?: 'text' | 'search'
+  /** Define the component display mode,such as which buttons should be visible */
+  displayMode?: 'clear-button' | 'search-and-clear-buttons' | 'search-button'
+  /** Define how the autocomplete component should be displayed. Possible values are: `overlay` (suggestions overlapping other components) and `container` (displays the suggestion within a container). */
+  containerMode: 'overlay' | 'container'
+}
+
+function SearchBar({
   placeholder,
   onInputChange,
   onGoToSearchPage,
@@ -46,21 +95,23 @@ const SearchBar = ({
   openAutocompleteOnFocus,
   blurOnSubmit,
   submitOnIconClick,
-  // eslint-disable-next-line react/prop-types
   displayMode,
   minSearchTermLength,
   autocompleteFullWidth,
   inputType,
-}) => {
+  containerMode = 'overlay',
+}: Props) {
   const intl = useIntl()
-  const container = useRef()
+  const container = useRef<HTMLDivElement>(null)
   const { navigate } = useRuntime()
   const handles = useCssHandles(CSS_HANDLES)
   const [searchTerm, setSearchTerm] = useState(inputValue)
-  const [inputErrorMessage, setInputErrorMessage] = useState()
+  const [inputErrorMessage, setInputErrorMessage] = useState<
+    string | undefined
+  >()
 
   const debouncedSetSearchTerm = useCallback(
-    debounce(newValue => {
+    debounce((newValue: string) => {
       setSearchTerm(newValue)
     }, SEARCH_DELAY_TIME),
     []
@@ -101,7 +152,7 @@ const SearchBar = ({
       }
 
       let page = 'store.product'
-      let params = {
+      let params: Record<string, string> = {
         slug: element.slug,
         id: element.productId,
       }
@@ -132,12 +183,12 @@ const SearchBar = ({
     return null
   }
 
-  const showInputErrorMessage = newInputErrorMessage => {
+  const showInputErrorMessage = (newInputErrorMessage: string) => {
     setInputErrorMessage(newInputErrorMessage)
   }
 
   const hideInputErrorMessage = () => {
-    setInputErrorMessage()
+    setInputErrorMessage(undefined)
   }
 
   const fallback = (
@@ -150,20 +201,31 @@ const SearchBar = ({
       iconBlockClass={iconBlockClass}
       inputErrorMessage={inputErrorMessage}
       onGoToSearchPage={onGoToSearchPage}
+      onClearInput={onClearInput}
     />
   )
 
   const isAutocompleteDeclared = Boolean(
-    useChildBlock({ id: AUTCOMPLETE_EXTENSION_ID })
+    useChildBlock({ id: AUTOCOMPLETE_EXTENSION_ID })
   )
 
   const SelectedAutocompleteResults = useMemo(() => {
-    return isAutocompleteDeclared
-      ? props => <ExtensionPoint id={AUTCOMPLETE_EXTENSION_ID} {...props} />
-      : props => <AutocompleteResults {...props} />
+    if (isAutocompleteDeclared) {
+      const AutoCompleteResultsWrapper = (props: AutocompleteResultsProps) => (
+        <ExtensionPoint id={AUTOCOMPLETE_EXTENSION_ID} {...props} />
+      )
+
+      return AutoCompleteResultsWrapper
+    }
+
+    const AutoCompleteResultsWrapper = (props: AutocompleteResultsProps) => (
+      <AutocompleteResults {...props} />
+    )
+
+    return AutoCompleteResultsWrapper
   }, [isAutocompleteDeclared])
 
-  const autocompleteContainerRef = useRef(null)
+  const autocompleteContainerRef = useRef<HTMLDivElement>(null)
 
   return (
     <div
@@ -200,13 +262,11 @@ const SearchBar = ({
                 // eslint-disable-next-line jsx-a11y/no-autofocus
                 autoFocus={autoFocus}
                 compactMode={compactMode}
-                onClearInput={onClearInput}
                 hasIconLeft={hasIconLeft}
                 iconClasses={iconClasses}
-                onGoToSearchPage={onGoToSearchPage}
+                openAutocompleteOnFocus={openAutocompleteOnFocus}
                 submitOnIconClick={submitOnIconClick}
                 displayMode={displayMode}
-                openAutocompleteOnFocus={openAutocompleteOnFocus}
                 inputType={inputType}
                 openMenu={openMenu}
                 inputErrorMessage={inputErrorMessage}
@@ -236,17 +296,18 @@ const SearchBar = ({
                   placeholder,
                   value: inputValue,
                   onChange: onInputChange,
+                  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+                  // @ts-expect-error this exists
                   onFocus: openAutocompleteOnFocus ? openMenu : undefined,
                 })}
+                onClearInput={onClearInput}
+                onGoToSearchPage={onGoToSearchPage}
               />
               <div ref={autocompleteContainerRef} />
-              <Overlay
-                fullWindow={autocompleteFullWidth}
-                alignment={autocompleteAlignment}
-                target={autocompleteContainerRef.current}
-              >
+
+              {containerMode === 'container' ? (
                 <SelectedAutocompleteResults
-                  parentContainer={container}
+                  parentContainer={container ?? undefined}
                   {...{
                     attemptPageTypeSearch,
                     isOpen,
@@ -260,58 +321,35 @@ const SearchBar = ({
                     customSearchPageUrl,
                   }}
                 />
-              </Overlay>
+              ) : (
+                <Overlay
+                  fullWindow={autocompleteFullWidth}
+                  alignment={autocompleteAlignment}
+                  target={autocompleteContainerRef.current ?? undefined}
+                >
+                  <SelectedAutocompleteResults
+                    parentContainer={container}
+                    {...{
+                      attemptPageTypeSearch,
+                      isOpen,
+                      getMenuProps,
+                      inputValue: searchTerm,
+                      getItemProps,
+                      selectedItem,
+                      highlightedIndex,
+                      closeMenu,
+                      onClearInput,
+                      customSearchPageUrl,
+                    }}
+                  />
+                </Overlay>
+              )}
             </div>
           )}
         </Downshift>
       </NoSSR>
     </div>
   )
-}
-
-SearchBar.propTypes = {
-  /** Placeholder to be used on the input */
-  placeholder: PropTypes.string.isRequired,
-  /** Current value of the input */
-  inputValue: PropTypes.string.isRequired,
-  /** Function to handle input changes */
-  onInputChange: PropTypes.func.isRequired,
-  /** Function to direct the user to the searchPage */
-  onGoToSearchPage: PropTypes.func.isRequired,
-  /** Function to clear the input */
-  onClearInput: PropTypes.func.isRequired,
-  /** Indentify when use the compact version of the component */
-  compactMode: PropTypes.bool,
-  /** Identify if the search icon is on left or right position */
-  hasIconLeft: PropTypes.bool,
-  /** Custom classes for the search icon */
-  iconClasses: PropTypes.string,
-  /** Block class for the search icon */
-  iconBlockClass: PropTypes.string,
-  /** Identify if the search input should autofocus or not */
-  autoFocus: PropTypes.bool,
-  /** Max width of the search bar */
-  maxWidth: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
-  /** A template for a custom url. It can have a substring ${term} used as placeholder to interpolate the searched term. (e.g. `/search?query=${term}`) */
-  customSearchPageUrl: PropTypes.string,
-  /** Uses the term the user has inputed to try to navigate to the proper
-   * page type (e.g. a department, a brand, a category)
-   */
-  attemptPageTypeSearch: PropTypes.bool,
-  /* Autocomplete Horizontal alignment */
-  autocompleteAlignment: PropTypes.string,
-  /** Identify if autocomplete should be open on input focus or not */
-  openAutocompleteOnFocus: PropTypes.bool,
-  /** Identify if input should blur on submit */
-  blurOnSubmit: PropTypes.bool,
-  /** Identify if icon should submit on click */
-  submitOnIconClick: PropTypes.bool,
-  /** Minimum search term length allowed */
-  minSearchTermLength: PropTypes.number,
-  /** If true, the autocomplete will fill the whole window horizontally */
-  autocompleteFullWidth: PropTypes.bool,
-  /** The type of the search input */
-  inputType: PropTypes.oneOf(['text', 'search']),
 }
 
 export default SearchBar