+++ react/typings/vtex.product-context.d.ts
@@ -25,6 +25,7 @@ interface Product {
   productName: string
   productReference: string
   brand: string
+  description: string
 }
 
 interface SkuSpecification {
@@ -89,5 +90,7 @@ declare module 'vtex.product-context' {
   }
 
   export const useProduct: () => ProductContext
+  type DispatchFunction = (payload: { type: string; args?: any }) => void
+  export const useProductDispatch: () => DispatchFunction
   export const ProductContext
 }