+++ react/typings/vtex.store-resources.d.ts
@@ -1,3 +1,5 @@
 declare module 'vtex.store-resources/OrderFormContext' {
   export const orderFormConsumer
 }
+
+declare module 'vtex.store-resources/QueryAddress'